Setup flyway, docker, mvn and tests

+++ src/main/resources/db/migration/V1_0_0__create_person_table.sql
@@ -1,6 +1,6 @@
 create table person
 (
-    id                    bigint       not null primary key,
+    id                    integer       not null primary key,
     name                  text,
-    age                   int
+    age                   integer
 );
